+++ openassessment/management/commands/create_oa_submissions_from_file.py
@@ -14,7 +14,7 @@ from opaque_keys.edx.keys import CourseKey
 import loremipsum
 from submissions import api as sub_api
 from openassessment.assessment.api import staff as staff_api
-from openassessment.function_imports import anonymous_id_for_user, modulestore
+from openassessment.runtime_imports.functions import anonymous_id_for_user, modulestore
 from openassessment.workflow import api as workflow_api
 from openassessment.xblock.data_conversion import create_rubric_dict
 from openassessment.staffgrader.models import SubmissionGradingLock
